requirements.txt
requests
pytest
pytest-html
allure-pytest
selenium
webdriver-manager